requirement.txt
flask
flask_bcrypt
flask_jwt_extended
flask_sqlalchemy
flask_migrate
mysql-connector-python